+++ esp-hal/src/gpio/placeholder.rs
@@ -85,25 +85,6 @@ impl crate::peripheral::Peripheral for NoPin {
 
 impl private::Sealed for NoPin {}
 
-impl embedded_hal_02::digital::v2::OutputPin for NoPin {
-    type Error = core::convert::Infallible;
-
-    fn set_high(&mut self) -> Result<(), Self::Error> {
-        Ok(())
-    }
-    fn set_low(&mut self) -> Result<(), Self::Error> {
-        Ok(())
-    }
-}
-impl embedded_hal_02::digital::v2::StatefulOutputPin for NoPin {
-    fn is_set_high(&self) -> Result<bool, Self::Error> {
-        Ok(false)
-    }
-    fn is_set_low(&self) -> Result<bool, Self::Error> {
-        Ok(false)
-    }
-}
-
 impl embedded_hal::digital::ErrorType for NoPin {
     type Error = core::convert::Infallible;
 }